scc: added defaults to the positioning
