+++ pycaption/scc/__init__.py
@@ -7,16 +7,18 @@ import string
 import textwrap
 
 from pycaption.base import (
-    BaseReader, BaseWriter, Caption, CaptionSet, CaptionNode,
+    BaseReader, BaseWriter, CaptionSet, CaptionNode,
 )
-from pycaption.geometry import Layout, Point, Size, UnitEnum
-from pycaption.exceptions import CaptionReadNoCaptions, CaptionReadSyntaxError
+from pycaption.exceptions import CaptionReadNoCaptions
 from .constants import (
     HEADER, COMMANDS, SPECIAL_CHARS, EXTENDED_CHARS, CHARACTERS,
     MICROSECONDS_PER_CODEWORD, CHARACTER_TO_CODE,
     SPECIAL_OR_EXTENDED_CHAR_TO_CODE, PAC_BYTES_TO_POSITIONING_MAP,
     PAC_HIGH_BYTE_BY_ROW, PAC_LOW_BYTE_BY_ROW_RESTRICTED,
 )
+from .specialized_collections import (
+    TimingCorrectingCaptionList, NotifyingDict, CaptionCreator,
+    InterpretableNodeCreator, DefaultProvidingPositionTracer)
 
 
 class SCCReader(BaseReader):
@@ -25,16 +27,16 @@ class SCCReader(BaseReader):
     This can be then later used for converting into any other supported formats
     """
     def __init__(self, *args, **kw):
-        self.caption_stash = _CaptionStash()
+        self.caption_stash = CaptionCreator()
         self.time_translator = _SccTimeTranslator()
 
         self.last_command = u''
 
-        self.buffer_dict = _NotifyingDict()
+        self.buffer_dict = NotifyingDict()
 
-        self.buffer_dict[u'pop'] = _InterpretableNodeStash()
-        self.buffer_dict[u'paint'] = _InterpretableNodeStash()
-        self.buffer_dict[u'roll'] = _InterpretableNodeStash()
+        self.buffer_dict[u'pop'] = InterpretableNodeCreator()
+        self.buffer_dict[u'paint'] = InterpretableNodeCreator()
+        self.buffer_dict[u'roll'] = InterpretableNodeCreator()
 
         # Call this method when the active key changes
         self.buffer_dict.add_change_observer(self._flush_implicit_buffers)
@@ -81,8 +83,12 @@ class SCCReader(BaseReader):
         if type(content) != unicode:
             raise RuntimeError(u'The content is not a unicode string.')
 
+        # Preparation. Clear the cached positioning from when processing
+        # other captions
+        DefaultProvidingPositionTracer.reset_default_positioning()
+
         self.simulate_roll_up = simulate_roll_up
-        self.time_translator.set_offset(offset * 1000000)
+        self.time_translator.offset = offset * 1000000
         # split lines
         lines = content.splitlines()
 
@@ -112,14 +118,14 @@ class SCCReader(BaseReader):
         # Calculate the end time from the current line
         time_translator = _SccTimeTranslator()
         time_translator.start_at(timing)
-        time_translator.set_offset(self.time_translator._offset)
+        time_translator.offset = self.time_translator.offset
 
         # But use the current time translator for the start time
         self.caption_stash.create_and_store(
             self.buffer, self.time_translator.get_time())
 
         self.caption_stash.correct_last_timing(time_translator.get_time())
-        self.buffer = _InterpretableNodeStash()
+        self.buffer = InterpretableNodeCreator()
 
     def _flush_implicit_buffers(self, old_key=None, *args):
         """Convert to Captions those buffers whose behavior is implicit.
@@ -230,7 +236,7 @@ class SCCReader(BaseReader):
                 self.caption_stash.create_and_store(
                     self.buffer, self.time
                 )
-                self.buffer = _InterpretableNodeStash()
+                self.buffer = InterpretableNodeCreator()
 
             self.time = self.time_translator.get_time()
 
@@ -250,7 +256,7 @@ class SCCReader(BaseReader):
             if not self.buffer.is_empty():
                 self.caption_stash.create_and_store(
                     self.buffer, self.time)
-                self.buffer = _InterpretableNodeStash()
+                self.buffer = InterpretableNodeCreator()
 
             # set rows to empty, configure start time for caption
             self.roll_rows = []
@@ -258,13 +264,13 @@ class SCCReader(BaseReader):
 
         # clear pop_on buffer
         elif word == u'94ae':
-            self.buffer = _InterpretableNodeStash()
+            self.buffer = InterpretableNodeCreator()
 
         # display pop_on buffer [End Of Caption]
         elif word == u'942f':
             self.time = self.time_translator.get_time()
             self.caption_stash.create_and_store(self.buffer, self.time)
-            self.buffer = _InterpretableNodeStash()
+            self.buffer = InterpretableNodeCreator()
 
         # roll up captions [Carriage Return]
         elif word == u'94ad':
@@ -284,7 +290,7 @@ class SCCReader(BaseReader):
             if not self.buffer_dict[u'paint'].is_empty():
                 self.caption_stash.create_and_store(
                     self.buffer_dict[u'paint'], self.time)
-                self.buffer = _InterpretableNodeStash()
+                self.buffer = InterpretableNodeCreator()
 
             # attempt to add proper end time to last caption(s)
             self.caption_stash.correct_last_timing(
@@ -331,11 +337,11 @@ class SCCReader(BaseReader):
                     self.roll_rows.pop(0)
 
                 self.roll_rows.append(self.buffer)
-                self.buffer = _InterpretableNodeStash.from_list(self.roll_rows)
+                self.buffer = InterpretableNodeCreator.from_list(self.roll_rows)
 
         # convert buffer and empty
         self.caption_stash.create_and_store(self.buffer, self.time)
-        self.buffer = _InterpretableNodeStash()
+        self.buffer = InterpretableNodeCreator()
 
         # configure time
         self.time = self.time_translator.get_time()
@@ -347,7 +353,7 @@ class SCCReader(BaseReader):
 class SCCWriter(BaseWriter):
 
     def __init__(self, *args, **kw):
-        pass
+        super(SCCWriter, self).__init__(*args, **kw)
 
     def write(self, caption_set):
         output = HEADER + u'\n\n'
@@ -390,25 +396,28 @@ class SCCWriter(BaseWriter):
         return output
 
     # Wrap lines at 32 chars
-    def _layout_line(self, caption):
+    @staticmethod
+    def _layout_line(caption):
         def caption_node_to_text(caption_node):
             if caption_node.type_ == CaptionNode.TEXT:
                 return unicode(caption_node.content)
             elif caption_node.type_ == CaptionNode.BREAK:
                 return u'\n'
-        caption_text = u''.join([caption_node_to_text(node)
-                                for node in caption.nodes])
+        caption_text = u''.join(
+            [caption_node_to_text(node) for node in caption.nodes])
         inner_lines = string.split(caption_text, u'\n')
         inner_lines_laid_out = [textwrap.fill(x, 32) for x in inner_lines]
         return u'\n'.join(inner_lines_laid_out)
 
-    def _maybe_align(self, code):
+    @staticmethod
+    def _maybe_align(code):
         # Finish a half-word with a no-op so we can move to a full word
         if len(code) % 5 == 2:
             code += u'80 '
         return code
 
-    def _maybe_space(self, code):
+    @staticmethod
+    def _maybe_space(code):
         if len(code) % 5 == 4:
             code += u' '
         return code
@@ -446,7 +455,8 @@ class SCCWriter(BaseWriter):
             code = self._maybe_align(code)
         return code
 
-    def _format_timestamp(self, microseconds):
+    @staticmethod
+    def _format_timestamp(microseconds):
         seconds_float = microseconds / 1000.0 / 1000.0
         # Convert to non-drop-frame timecode
         seconds_float *= 1000.0 / 1001.0
@@ -460,209 +470,14 @@ class SCCWriter(BaseWriter):
         return u'%02d:%02d:%02d:%02d' % (hours, minutes, seconds, frames)
 
 
-class _TimingCorrectingCaptionList(list):
-    """List of captions. Will know to correct the last caption's end time
-    when adding a new caption.
-
-    Also, doesn't allow Nones or empty captions
-    """
-    def append(self, p_object):
-        """When appending a new caption to the list, make sure the last one
-        has an end. Also, don't add empty captions
-
-        :type p_object: Caption
-        """
-        if p_object is None:
-            return
-
-        if len(self) > 0 and self[-1].end == 0:
-            self[-1].end = p_object.start
-
-        if p_object.nodes:
-            super(_TimingCorrectingCaptionList, self).append(p_object)
-
-    def extend(self, iterable):
-        """Adds the elements in the iterable to the list
-
-        :param iterable: any iterable
-        """
-        for elem in iterable:
-            self.append(elem)
-
-
-class _CaptionStash(object):
-    """Creates and maintains a collection of Captions
-    """
-    def __init__(self):
-        self._collection = _TimingCorrectingCaptionList()
-
-        # subset of self._collection;
-        # captions here will be susceptible to time corrections
-        self._still_editing = []
-
-    def correct_last_timing(self, end_time, force=False):
-        """Called to set the time on the last Caption(s) stored with no end
-        time
-
-        :type force: bool
-        :param force: Set the end time even if there's already an end time
-
-        :type end_time: int
-        :param end_time: microseconds; the end of the caption;
-        """
-        if not self._still_editing:
-            return
-
-        if force:
-            captions_to_correct = self._still_editing
-        else:
-            captions_to_correct = (
-                caption for caption in self._still_editing
-                if caption.end == 0
-            )
-
-        for caption in captions_to_correct:
-            caption.end = end_time
-
-    def get_last(self):
-        """Returns the last caption stored (for setting the time on it),
-        or None
-
-        :rtype: Caption
-        """
-        if len(self._collection) > 0:
-            return self._collection[-1]
-
-        return None
-
-    def create_and_store(self, node_buffer, start):
-        """Interpreter method, will convert the buffer into one or more Caption
-        objects, storing them internally.
-
-        :type node_buffer: _InterpretableNodeStash
-
-        :type start: int
-        :param start: the start time in microseconds
-        """
-        if node_buffer.is_empty():
-            return
-
-        caption = Caption()
-        caption.start = start
-        caption.end = 0  # Not yet known; filled in later
-        self._still_editing = [caption]
-
-        open_italic = False
-
-        for element in node_buffer:
-            # skip empty elements
-            if element.is_empty():
-                continue
-
-            elif element.requires_repositioning():
-                self._remove_extra_italics(caption)
-                open_italic = False
-                caption = Caption()
-                caption.start = start
-                caption.end = 0
-                self._still_editing.append(caption)
-
-            # handle line breaks
-            elif element.is_explicit_break():
-                new_nodes = self._translate_break(open_italic)
-                open_italic = False
-                caption.nodes.extend(new_nodes)
-
-            # handle open italics
-            elif element.sets_italics_on():
-                # add italics
-                caption.nodes.append(
-                    CaptionNode.create_style(True, {u'italics': True}))
-                # open italics, no longer first element
-                open_italic = True
-
-            # handle clone italics
-            elif element.sets_italics_off() and open_italic:
-                caption.nodes.append(
-                    CaptionNode.create_style(False, {u'italics': True}))
-                open_italic = False
-
-            # handle text
-            else:
-                # add text
-                layout_info = _get_layout_from_tuple(element.position)
-                caption.nodes.append(
-                    CaptionNode.create_text(
-                        element.get_text(), layout_info=layout_info),
-                )
-                caption.layout_info = layout_info
-
-        # close any open italics left over
-        if open_italic:
-            caption.nodes.append(
-                CaptionNode.create_style(False, {u'italics': True}))
-
-        # remove extraneous italics tags in the same caption
-        self._remove_extra_italics(caption)
-
-        self._collection.extend(self._still_editing)
-
-    @staticmethod
-    def _translate_break(open_italic):
-        """Depending on the context, translates a line break into one or more
-        nodes, returning them.
-
-        :param open_italic: bool
-        :rtype: tuple
-        """
-        new_nodes = []
-
-        if open_italic:
-            new_nodes.append(CaptionNode.create_style(
-                False, {u'italics': True}))
-
-        # add line break
-        new_nodes.append(CaptionNode.create_break())
-
-        return new_nodes
-
-    @staticmethod
-    def _remove_extra_italics(caption):
-        """Legacy logic slightly refactored. Removes STYLE nodes that would
-        surround a BREAK node.
-
-        See CaptionNode
-
-        :type caption: Caption
-        """
-        i = 0
-        length = max(0, len(caption.nodes) - 2)
-        while i < length:
-            if (caption.nodes[i].type_ == CaptionNode.STYLE and
-                    caption.nodes[i].content[u'italics'] and
-                    caption.nodes[i + 1].type_ == CaptionNode.BREAK and
-                    caption.nodes[i + 2].type_ == CaptionNode.STYLE and
-                    caption.nodes[i + 2].content[u'italics']):
-                # Remove the two italics style nodes
-                caption.nodes.pop(i)
-                caption.nodes.pop(i + 1)
-                length -= 2
-            i += 1
-
-    def get_all(self):
-        """Returns the Caption collection as a list
-
-        :rtype: list
-        """
-        return list(self._collection)
-
-
 class _SccTimeTranslator(object):
     """Converts SCC time to microseconds, keeping track of frames passed
     """
     def __init__(self):
         self._time = 0
-        self._offset = 0
+
+        # microseconds. The offset from which we begin the time calculation
+        self.offset = 0
         self._frames = 0
 
     def get_time(self):
@@ -673,7 +488,7 @@ class _SccTimeTranslator(object):
         """
         return self._translate_time(
             self._time[:-2] + unicode(int(self._time[-2:]) + self._frames),
-            self._offset
+            self.offset
         )
 
     @staticmethod
@@ -716,442 +531,10 @@ class _SccTimeTranslator(object):
         self._time = timespec
         self._frames = 0
 
-    def set_offset(self, offset):
-        """Sets the offset from which to calculate the time
-
-        :param offset: number of microseconds. will be deducted when
-            calculating the time
-        :type offset: int
-        """
-        self._offset = offset
-
     def increment_frames(self):
         """After a command was processed, we'd increment the number of frames
         """
         self._frames += 1
-
-
-class _InterpretableNodeStash(object):
-    """Creates _InterpretableNode instances from characters and commands,
-    and stores them internally in a buffer.
-    """
-    def __init__(self, collection=None):
-        if not collection:
-            self._collection = []
-        else:
-            self._collection = collection
-        self._position_tracer = _PositioningTracer()
-
-    def is_empty(self):
-        """Whether any text was added to the buffer
-        """
-        return not any(element.text for element in self._collection)
-
-    def discard_last_char(self):
-        """This was previously used to discard a mid-row command, but generated
-        a bug, by confusing special/extended characters with mid-row commands.
-        This method should only discard the last character if it's a mid-row
-        command, and it's equal to the current command - that however should
-        be handled by _handle_double_command already
-        """
-        pass
-
-    def add_chars(self, *chars):
-        """Adds characters to a text node (last text node, or a new one)
-
-        :param chars: tuple containing text (unicode)
-        """
-        if not chars:
-            return
-
-        current_position = self._position_tracer.get_current_position()
-
-        # get or create a usable node
-        if self._collection:
-            node = self._collection[-1]
-        else:
-            # create first node
-            node = _InterpretableNode(position=current_position)
-            self._collection.append(node)
-
-        # handle a simple line break
-        if self._position_tracer.is_linebreak_required():
-            # must insert a line break here
-            self._collection.append(_InterpretableNode.create_break(
-                position=current_position))
-            node = _InterpretableNode.create_text(current_position)
-            self._collection.append(node)
-            self._position_tracer.acknowledge_linebreak_consumed()
-
-        # handle completely new positioning
-        elif self._position_tracer.is_repositioning_required():
-            # this node will have a different positioning than the previous one
-            self._collection.append(
-                _InterpretableNode.create_repositioning_command())
-            node = _InterpretableNode.create_text(current_position)
-            self._collection.append(node)
-            self._position_tracer.acknowledge_position_changed()
-
-        node.add_chars(*chars)
-
-    def interpret_command(self, command):
-        """Given a command determines whether tu turn italics on or off,
-        or to set the positioning
-
-        This is mostly used to convert from the legacy-style commands
-
-        :type command: unicode
-        """
-        self._update_positioning(command)
-
-        text = COMMANDS.get(command, u'')
-
-        if u'<$>{italic}<$>' in text:
-            self._collection.append(
-                _InterpretableNode.create_italics_style(
-                    self._position_tracer.get_current_position())
-            )
-        elif u'<$>{end-italic}<$>' in text:
-            self._collection.append(
-                _InterpretableNode.create_italics_style(
-                    self._position_tracer.get_current_position(),
-                    turn_on=False
-                )
-            )
-
-    def _update_positioning(self, command):
-        """Sets the positioning information to use for the next nodes
-
-        :type command: unicode
-        """
-        if len(command) != 4:
-            return
-
-        first, second = command[:2], command[2:]
-
-        try:
-            positioning = PAC_BYTES_TO_POSITIONING_MAP[first][second]
-        except KeyError:
-            pass
-        else:
-            self._position_tracer.update_positioning(positioning)
-
-    def __iter__(self):
-        return iter(self._collection)
-
-    @classmethod
-    def from_list(cls, stash_list):
-        """Having received a list of instances of this class, creates a new
-        instance that contains all the nodes of the previous instances
-        (basically concatenates the many stashes into one)
-
-        :param stash_list: a list of instances of this class
-        :type stash_list: list[_InterpretableNodeStash]
-        :rtype: _InterpretableNodeStash
-        """
-        instance = cls()
-        new_collection = instance._collection
-
-        for idx, stash in enumerate(stash_list):
-            new_collection.extend(stash._collection)
-
-            # use space to separate the stashes, but don't add final space
-            if idx < len(stash_list) - 1:
-                try:
-                    instance._collection[-1].add_chars(u' ')
-                except AttributeError:
-                    pass
-
-        return instance
-
-
-class _PositioningTracer(object):
-    """Helps determine the positioning of a node, having kept track of
-    positioning-related commands.
-
-    Acts like a state-machine, with 2
-
-    """
-    def __init__(self, positioning=None):
-        """
-        :param positioning: positioning information (row, column)
-        :type positioning: tuple[int]
-        """
-        self._positions = [positioning]
-        self._break_required = False
-        self._repositioning_required = False
-
-    def update_positioning(self, positioning):
-        """Being notified of a position change, updates the internal state,
-        to as to be able to tell if it was a trivial change (a simple line
-        break) or not.
-
-        :type positioning: tuple[int]
-        :param positioning: a tuple (row, col)
-        """
-        current = self._positions[-1]
-
-        if not current:
-            if positioning:
-                # set the positioning for the first time
-                self._positions = [positioning]
-            return
-
-        row, col = current
-        new_row, _ = positioning
-
-        # is the new position simply one line below?
-        if new_row == row + 1:
-            self._positions.append((new_row, col))
-            self._break_required = True
-        else:
-            # reset the "current" position altogether.
-            self._positions = [positioning]
-            self._repositioning_required = True
-
-    def get_current_position(self):
-        """Returns the current usable position
-
-        :rtype: tuple[int]
-        """
-        if not any(self._positions):
-            raise CaptionReadSyntaxError(
-                u'No Preamble Address Code [PAC] was provided'
-            )
-        else:
-            return self._positions[0]
-
-    def is_repositioning_required(self):
-        """Determines whether the current positioning has changed non-trivially
-
-        Trivial would be mean that a line break should suffice.
-        :rtype: bool
-        """
-        return self._repositioning_required
-
-    def acknowledge_position_changed(self):
-        """Acknowledge the position tracer that the position was changed
-        """
-        self._repositioning_required = False
-
-    def is_linebreak_required(self):
-        """If the current position is simply one line below the previous.
-        :rtype: bool
-        """
-        return self._break_required
-
-    def acknowledge_linebreak_consumed(self):
-        """Call to acknowledge that the line required was consumed
-        """
-        self._break_required = False
-
-
-class _InterpretableNode(object):
-    """Value object, that can contain text information, or interpretable
-    commands (such as explicit line breaks or turning italics on/off)
-    """
-    TEXT = 0
-    BREAK = 1
-    ITALICS_ON = 2
-    ITALICS_OFF = 3
-    CHANGE_POSITION = 4
-
-    def __init__(self, text=None, position=None, type_=0):
-        """
-        :type text: unicode
-        :param position: a tuple of ints (row, column)
-        :param type_: self.TEXT | self.BREAK | self.ITALICS
-        :type type_: int
-        """
-        self.text = text
-        self.position = position
-        self._type = type_
-
-    def add_chars(self, *args):
-        """This being a text node, add characters to it.
-        :param args:
-        :type args: tuple[unicode]
-        :return:
-        """
-        if self.text is None:
-            self.text = u''
-
-        self.text += u''.join(args)
-
-    def is_text_node(self):
-        """
-        :rtype: bool
-        """
-        return self._type == self.TEXT
-
-    def is_empty(self):
-        """
-        :rtype: bool
-        """
-        if self._type == self.TEXT:
-            return not self.text
-
-        return False
-
-    def is_explicit_break(self):
-        """
-        :rtype: bool
-        """
-        return self._type == self.BREAK
-
-    def sets_italics_on(self):
-        """
-        :rtype: bool
-        """
-        return self._type == self.ITALICS_ON
-
-    def sets_italics_off(self):
-        """
-        :rtype: bool
-        """
-        return self._type == self.ITALICS_OFF
-
-    def requires_repositioning(self):
-        """Whether the node must be interpreted as a change in positioning
-
-        :rtype: bool
-        """
-        return self._type == self.CHANGE_POSITION
-
-    def get_text(self):
-        """A little legacy code.
-        """
-        return u' '.join(self.text.split())
-
-    @classmethod
-    def create_break(cls, position):
-        """Create a node, interpretable as an explicit line break
-
-        :type position: tuple[int]
-        :param position: a tuple (row, col) containing the positioning info
-
-        :rtype: _InterpretableNode
-        """
-        return cls(type_=cls.BREAK, position=position)
-
-    @classmethod
-    def create_text(cls, position, *chars):
-        """Create a node interpretable as text
-
-        :type position: tuple[int]
-        :param position: a tuple (row, col) to mark the positioning
-
-        :type chars: tuple[unicode]
-        :param chars: characters to add to the text
-
-        :rtype: _InterpretableNode
-        """
-        return cls(u''.join(chars), position=position)
-
-    @classmethod
-    def create_italics_style(cls, position, turn_on=True):
-        """Create a node, interpretable as a command to switch italics on/off
-
-        :type position: tuple[int]
-        :param position: a tuple (row, col) to mark the positioning
-
-        :type turn_on: bool
-        :param turn_on: whether to turn the italics on or off
-
-        :rtype: _InterpretableNode
-        """
-        return cls(
-            position=position,
-            type_=cls.ITALICS_ON if turn_on else cls.ITALICS_OFF
-        )
-
-    @classmethod
-    def create_repositioning_command(cls):
-        """Create node interpretable as a command to change the current
-        position
-        """
-        return cls(type_=cls.CHANGE_POSITION)
-
-    def __repr__(self):
-        if self._type == self.BREAK:
-            return u'<INode: BR>'
-        elif self._type == self.TEXT:
-            return u'<INode: "{}">'.format(self.text)
-        elif self._type in (self.ITALICS_ON, self.ITALICS_OFF):
-            return u'<INode: italics {}>'.format(
-                u'on' if self._type == self.ITALICS_ON else u'off')
-        else:
-            return u'<INode: change position>'
-
-
-class _NotifyingDict(dict):
-    """Dictionary-like object, that treats one key as 'active',
-    and notifies observers if the active key changed
-    """
-    # need an unhashable object as initial value for the active key
-    _guard = {}
-
-    def __init__(self, *args, **kwargs):
-        super(_NotifyingDict, self).__init__(*args, **kwargs)
-        self.active_key = self._guard
-        self.observers = []
-
-    def set_active(self, key):
-        """Sets the active key
-
-        :param key: any hashable object
-        """
-        if key not in self:
-            raise ValueError(u'No such key present')
-
-        # Notify observers of the change
-        if key != self.active_key:
-            for observer in self.observers:
-                observer(self.active_key, key)
-
-        self.active_key = key
-
-    def get_active(self):
-        """Returns the value corresponding to the active key
-        """
-        if self.active_key is self._guard:
-            raise KeyError(u'No active key set')
-
-        return self[self.active_key]
-
-    def add_change_observer(self, observer):
-        """Receives a callable function, which it will call if the active
-        element changes.
-
-        The observer will receive 2 positional arguments: the old and new key
-
-        :param observer: any callable that can be called with 2 positional
-            arguments
-        """
-        if not callable(observer):
-            raise TypeError(u'The observer should be callable')
-
-        self.observers.append(observer)
-
-
-def _get_layout_from_tuple(position_tuple):
-    """Create a Layout object from the positioning information given
-
-    The row can have a value from 1 to 15 inclusive. (vertical positioning)
-    The column can have a value from 0 to 31 inclusive. (horizontal)
-
-    :param position_tuple: a tuple of ints (row, col)
-    :type position_tuple: tuple
-    :rtype: Layout
-    """
-    if not position_tuple:
-        return None
-
-    row, column = position_tuple
-
-    horizontal = Size(100 * column / 32.0, UnitEnum.PERCENT)
-    vertical = Size(100 * (row - 1) / 15.0, UnitEnum.PERCENT)
-    return Layout(origin=Point(horizontal, vertical))
 
 
 def _is_pac_command(word):
